Create 01-Introduction-to-Python-ALICE
Revamping of introductory Python course material. I have significantly reduced the amount of material that is covered, with the aim to practice and master these concepts. This material covers: variables, data types, data structures (lists and indexing), control flow with if statements and loops.

I have created a complementary worksheet along with this notebook. The worksheet is linked at the top of the notebook.

Material is targeted towards people who have no prior coding experience in any language.